Language names available on alternate link tags.
### What
Added an attribute `data-lang-name` which can contain the language name for each alternate link.

### Why
So that the developer can not only get the alternate languages, but they can get the name of each alternate language as defined by them. I considered adding a list of names to the library, but that would be unnecessary large, at about 1MB worth of JavaScript.